requirements.txt
# Django
Django==2.0
djangorestframework==3.7.7
django-rest-swagger==2.1.2  # API Documentation
djangorestframework-jwt==1.11.0  # JSON Web Token for DRF
django-adminlte2==0.1.6  # Beautiful Admin Page
django-extensions==1.9.6  # More command for manage.py: reset_db
django-filter==2.0.0.dev1
gunicorn==19.7.1

# Database
psycopg2==2.7.3.1
redis==2.10.6
django-redis==4.8.0

# Automation Deployment

# Other packages
requests==2.18.4
Pillow==4.3  # Upload image
ipython==6.2.1
whitenoise==3.3.1 # Serve static files
html-json-forms==1.0.0django-adminlte2==0.1.6
